Move web workspace package screens
