+++ apps/web/src/screens/settings/workspace/packages/WorkspaceImportScreen.tsx
@@ -2,17 +2,17 @@ import { useEffect, useRef, useState, type ReactElement } from "react";
 import {
   confirmWorkspacePackageImport,
   previewWorkspacePackageImport,
-} from "../../../api";
-import { useAppData } from "../../../appData";
-import { requireCloudInstallationId } from "../../../appData/sync/local/syncCloudSettings";
-import { useAppErrorDialog } from "../../../appError/AppErrorContext";
-import { useI18n } from "../../../i18n";
-import { captureAppOperationError } from "../../../observability/appOperationObservation";
+} from "../../../../api";
+import { useAppData } from "../../../../appData";
+import { requireCloudInstallationId } from "../../../../appData/sync/local/syncCloudSettings";
+import { useAppErrorDialog } from "../../../../appError/AppErrorContext";
+import { useI18n } from "../../../../i18n";
+import { captureAppOperationError } from "../../../../observability/appOperationObservation";
 import type {
   WorkspacePackageImportConfirmOptions,
   WorkspacePackageImportPreviewResponse,
-} from "../../../types";
-import { SettingsShell } from "../SettingsShared";
+} from "../../../../types";
+import { SettingsShell } from "../../SettingsShared";
 
 type PackageMetadataRow = Readonly<{
   label: string;
